requirements.txt
streamlit
openai>=1.12.0
python-dotenv
neomodel
numpy
networkx
matplotlib
pandas
requests
instructor
pydantic>=2.0.0
typing-extensions